\name{stlda-package}
\alias{stlda-package}
\alias{stlda}
\docType{package}
\title{
\packageTitle{stlda}
}
\description{
\packageDescription{stlda}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stlda}
\packageIndices{stlda}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stlda}

Maintainer: \packageMaintainer{stlda}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
