% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_funcs.R
\name{stlda_em}
\alias{stlda_em}
\title{SpatioTemporal LDA estimation via Expecation Maximization}
\usage{
stlda_em(docs, K, V, THETAs_init, OMEGAs_init, PHI_init, max_iters = 300,
  thresh = 0.001, eta = rep(1, V), sigma_om = 1, sigma_theta = 1,
  sigma_b = 1, l_kern = 1)
}
\arguments{
\item{docs}{docs}

\item{K}{K}

\item{V}{V}

\item{THETAs_init}{Initial value for document-topic prevalences. Initial values are important due to the nonconvex nature of the posterior.}

\item{OMEGAs_init}{Initial value for space-time-topic in R^d.}

\item{PHI_init}{Initial value for topic-word matrix}

\item{max_iters}{Max outer EM iterations}

\item{thresh}{Mean l2 norm on OMEGA matrices to stop iteration.}

\item{eta}{A vector of length V, the prior on PHI}

\item{sigma_om}{The variance for OMEGA, for now fixed}

\item{sigma_theta}{Random effect variance for each document}

\item{sigma_b}{Covariance for spatial kernel.}

\item{l_kern}{Lengthscale for spatial kernel.}
}
\description{
SpatioTemporal LDA estimation via Expecation Maximization
}
